Under_100_WebTools/Get_content_from_locked_page/requirements.txt
requests
beautifulsoup4
python-docx